YUK-13056 RFN OV/UV config does not show In Sync unless read manually - Fix cmd RfnSetOvUvSetOverVoltageThresholdCommand and RfnSetOvUvSetUnderVoltageThresholdCommand to threshold has double instead of unsigned long - Added validate RfnSetOvUvSetThresholdCommand of threshold not exceeding precision of 0.001 - Fix RfnResidentialDevice execute install to check if dynamic pao value exist - Added unit test to verify pao dynamic data for Tou configuration, ovuv configuration and voltage profile

+++ yukon-server/RTDB/include/cmd_rfn_OvUvConfiguration.h
@@ -156,7 +156,7 @@ struct IM_EX_DEVDB RfnSetOvUvSetOverVoltageThresholdCommand : public RfnSetOvUvS
 
     virtual void invokeResultHandler( ResultHandler &rh ) const;
 
-    const unsigned ovThresholdValue;
+    const double ovThresholdValue;
 };
 
 struct IM_EX_DEVDB RfnSetOvUvSetUnderVoltageThresholdCommand : public RfnSetOvUvSetThresholdCommand
@@ -165,7 +165,7 @@ struct IM_EX_DEVDB RfnSetOvUvSetUnderVoltageThresholdCommand : public RfnSetOvUv
 
     virtual void invokeResultHandler( ResultHandler &rh ) const;
 
-    const unsigned uvThresholdValue;
+    const double uvThresholdValue;
 };
